+++ test/dom/index.js
@@ -249,3 +249,15 @@ it('should only update if shouldUpdate returns true', function () {
   })
 });
 
+it('should not allow setting the state during render', function (done) {
+  var Impure = component(function(){
+    this.setState({ foo: 'bar' });
+    return dom();
+  });
+  try {
+    var app = scene(Impure)
+    mount(app)
+  } catch(e) {
+    return done();
+  }
+});
