+++ test/dom/replace.js
