+++ test/index.js
@@ -5,14 +5,13 @@ describe('component', function(){
 
 describe('render', function () {
   require('./dom')
-  require('./dom/render-hook')
   require('./dom/mount-hook')
   require('./dom/update-hook')
   require('./dom/props')
   require('./dom/elements')
   require('./dom/text')
   require('./dom/attributes')
-  require('./dom/update-replace')
+  require('./dom/replace')
   require('./dom/events')
   require('./dom/state')
 })
